[refactor] #13 exception 패키지 위치 변경
